threading

+++ sd_app/views.py
@@ -157,9 +157,11 @@ def delete_search():
     idx = data['idx']
     search = Search.query.get(idx)
 
-    filepath = os.path.join(views.root_path, 'model_outputs', search.csv_path)
-    print(f"File path to be deleted: {filepath}")
-    os.remove(filepath)
+    if search.csv_path != "In progress":
+        filepath = os.path.join(views.root_path, 'model_outputs', search.csv_path)
+        if os.path.exists(filepath):
+            print(f"File path to be deleted: {filepath}")
+            os.remove(filepath)
 
     db.session.delete(search)
     db.session.commit()
